broke out 'linebreakTypesExcept' helper function to reduce size of the main config file
